\name{wRappoveRpass-package}
\alias{wRappoveRpass-package}
\alias{wRappoveRpass}
\docType{package}
\title{
\packageTitle{wRappoveRpass}
}
\description{
\packageDescription{wRappoveRpass}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{wRappoveRpass}
\packageIndices{wRappoveRpass}
A simple wrapper for OSM Overpass API and few utils functions.
At the moment the library has two functions: a wrapper and a function that turns polygons into points.
}
\author{
\packageAuthor{wRappoveRpass}

Maintainer: \packageMaintainer{wRappoveRpass}
}
